// src/app/interfaces/pago-print.interface.ts
export interface PagoPrintInterface {
    id:                    number;
    tipo_Cargo_Abono:      number;
    monto:                 number;
    tipo_Cambio:           number;
    moneda:                number;
    cambio:                number;
    monto_Moneda:          number;
    referencia:            string;
    autorizacion:          string;
    fDes_Tipo_Cargo_Abono: string;
}